+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @verdant-web/cli
 
+## 4.4.0
+
+### Minor Changes
+
+- c701cfd: Add React hook "useOnChange" (fires a callback on entity changes). Add "deep" option to React hook "useWatch" to watch for deep changes. Remove the "field" variant of "useWatch." CLI must be updated to correctly generate new React hook types.
+
 ## 4.3.2
 
 ### Patch Changes